test: working integration test
