--- tests/unit/configs/integration_test.yaml
@@ -1,6 +1,0 @@
-
-            checkpoint: checkpoint.pth
-            input:
-                grib: input.grib
-            output:
-                grib: output.grib
